Moved scraping to data_containers, removed script from utils. Updated env files
