+++ environment_python_3_8.yml
@@ -23,6 +23,7 @@ channels:
 dependencies:
   - astropy=4.3.1
   - bokeh=2.4.2
+  - beautifulsoup4=4.9.3
   - codecov=2.1.11
   - cryptography=36.0.0
   - django=3.1.7
