+++ requirements.txt
@@ -2,6 +2,7 @@ asdf==2.8.3
 astropy==4.3.1
 astroquery==0.4.5
 bandit==1.7.1
+beautifulsoup4==4.9.3
 bokeh==2.4.2
 codecov==2.1.12
 crds==11.5.2
